rag-qa-system/requirements.txt
# Core
sentence-transformers
faiss-cpu
# Core
torch
transformers
sentence-transformers
faiss-cpu
numpy
scikit-learn

# Ingestion
pymupdf
nltk
pypdf

# App
streamlit

# Utilities
tqdm
python-dotenv
pandas
openai
